Add the sendCardData redux thunk function
