+++ src/App.js
@@ -4,6 +4,7 @@ import "./App.css";
 import Auth from "./components/Auth";
 import Layout from "./components/Layout";
 import Notification from "./components/Notification";
+import { sendCartData } from "./store/cart-slice";
 import { uiActions } from "./store/ui-slice";
 
 function App() {
@@ -20,41 +21,8 @@ function App() {
       setFirstRender(false);
       return;
     }
-    const sendRequest = async () => {
-      //Send state as sending request
-      dispatch(uiActions.showNotification({
-        message: "Sending Request",
-        type: "warning",
-        open: true
-      }));
-      
-      const res = await fetch("https://redux-http-17d55-default-rtdb.firebaseio.com/cartItems.json",
-        {
-        method: "PUT",
-        body: JSON.stringify(cart)
-        }
-      );
-      const data = await res.json();
-
-
-      //Send state as request is successful
-      dispatch(uiActions.showNotification({
-        open: true,
-        message: "Successfully sent request to the Database",
-        type: "success"
-      }));
-
-    };
-    sendRequest().catch(err => {
-      //Send state as error
-      dispatch(uiActions.showNotification({
-        open: true,
-        message: "Sending request failed",
-        type: "error"
-      }));
-    });
-  }, [cart]);
-
+    dispatch(sendCartData(cart));
+  }, [cart, dispatch, notification, isFirstRender]);
 
   return (
     <div className="App">
